+++ app/src/main/java/com/example/appfrisaahorasi/pantallas/InicioSesion/PrimerInicioSesion.kt
@@ -1,4 +1,4 @@
-package com.example.appfrisaahorasi.pantallas.Registro.Usuario
+package com.example.appfrisaahorasi.pantallas.InicioSesion
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -36,7 +36,7 @@ import androidx.compose.runtime.setValue
 
 @Preview
 @Composable
-fun registroU2() {
+fun PrimerInicioSesion() {
 
     val tags: List<String> = listOf(
         "Ambientalistas",
