+++ app/src/main/java/com/example/appfrisaahorasi/pantallas/Registro/Organizacion/registroSC2.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.example.appfrisaahorasi.pantallas.Registro.RegistroViewModel
 
 class RegistroSC2Activity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
