+++ app/src/main/java/com/example/appfrisaahorasi/pantallas/Registro/Organizacion/registroSC5.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.example.appfrisaahorasi.pantallas.Registro.Usuario.ClickableBox
+import com.example.appfrisaahorasi.pantallas.InicioSesion.ClickableBox
 
 @Preview
 @Composable
