+++ app/src/main/java/com/example/appfrisaahorasi/pantallas/Registro/Usuario/registroU.kt
@@ -1,9 +1,6 @@
 package com.example.appfrisaahorasi.pantallas.Registro.Usuario
 
 import android.annotation.SuppressLint
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
@@ -13,7 +10,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
@@ -25,23 +21,19 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.navigation.NavController
 import com.example.appfrisaahorasi.pantallas.Registro.Organizacion.CustomTextField
 import com.example.appfrisaahorasi.pantallas.Registro.Organizacion.PasswordTextField
-import com.example.appfrisaahorasi.pantallas.Registro.Organizacion.RegistroViewModel
+import com.example.appfrisaahorasi.pantallas.Registro.RegistroViewModel
 
-class RegistroUActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            RegistroUScreen()
-        }
-    }
-}
 
-@Preview
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
-fun RegistroUScreen() {
+fun RegistroUScreen(
+    navController: NavController,
+    viewModel: RegistroViewModel = androidx.lifecycle.viewmodel.compose.viewModel()
+    )
+    {
     val viewModel: RegistroViewModel = viewModel()
 
     Scaffold( // nav bar iría aquí
@@ -65,34 +57,24 @@ fun RegistroUScreen() {
                     fontWeight = FontWeight.Bold,
                     color = Color.Black
                 )
-
                 Text(text = "Ingrese sus datos en los siguientes campos",
                     modifier = Modifier.padding(15.dp),
                     textAlign = TextAlign.Center,
                     color = Color.DarkGray,
                     fontSize = 15.sp)
-
-
-
                 Text(
-                    text = "Obligatorio *",
+                    text = "NOMBRE Y APELLIDOS *",
                     textAlign = TextAlign.Start,
                     fontSize = 12.sp,
                     textDecoration = TextDecoration.None,
                     letterSpacing = 0.sp,
-
                     overflow = TextOverflow.Ellipsis,
                     modifier = Modifier
                         .align(alignment = Alignment.Start)
                         .padding(18.dp, 6.dp, 18.dp)
-
-                        .width(77.dp)
-
+                        .width(207.dp)
                         //.height(18.dp)
-
                         .alpha(1f),
-
-
                     fontWeight = FontWeight.Medium,
                     fontStyle = FontStyle.Normal,
                     color = Color.Gray
@@ -101,38 +83,31 @@ fun RegistroUScreen() {
                 CustomTextField(
                     value = viewModel.nombre,
                     onValueChange = { newValue -> viewModel.onNombreChanged(newValue)},
-                    placeholder = "Nombre",
+                    placeholder = " ",
                     keyboardType = KeyboardType.Text
                 )
 
                 Text(
-                    text = "Obligatorio *",
+                    text = "MEDIO DE CONTACTO",
                     textAlign = TextAlign.Start,
                     fontSize = 12.sp,
                     textDecoration = TextDecoration.None,
                     letterSpacing = 0.sp,
-
                     overflow = TextOverflow.Ellipsis,
                     modifier = Modifier
                         .align(alignment = Alignment.Start)
                         .padding(18.dp, 6.dp, 18.dp)
-
-                        .width(77.dp)
-
+                        .width(150.dp)
                         //.height(18.dp)
-
                         .alpha(1f),
-
-
                     fontWeight = FontWeight.Medium,
                     fontStyle = FontStyle.Normal,
                     color = Color.Gray
                 )
-
                 CustomTextField(
                     value = viewModel.celular,
                     onValueChange = { newValue -> viewModel.onCelularChanged(newValue) },
-                    placeholder = "Celular",
+                    placeholder = "Celular *",
                     keyboardType = KeyboardType.Phone
                 )
 
@@ -145,7 +120,7 @@ fun RegistroUScreen() {
 
                 var passwordHidden by rememberSaveable { mutableStateOf(true) }
                 Text(
-                    text = "Obligatorio *",
+                    text = "CONTRASEÑA *",
                     textAlign = TextAlign.Start,
                     fontSize = 12.sp,
                     textDecoration = TextDecoration.None,
@@ -155,14 +130,9 @@ fun RegistroUScreen() {
                     modifier = Modifier
                         .align(alignment = Alignment.Start)
                         .padding(18.dp, 6.dp, 18.dp)
-
-                        .width(77.dp)
-
+                        .width(100.dp)
                         //.height(18.dp)
-
                         .alpha(1f),
-
-
                     fontWeight = FontWeight.Medium,
                     fontStyle = FontStyle.Normal,
                     color = Color.Gray
@@ -170,45 +140,14 @@ fun RegistroUScreen() {
                 PasswordTextField(
                     value = viewModel.contrasena,
                     onValueChange = { newValue -> viewModel.onContrasenaChanged(newValue) },
-                    placeholder = "Contraseña",
+                    placeholder = "Minimo 6 caracteres",
                     keyboardType = KeyboardType.Password,
                     isPassword = true,
                     passwordVisibility = passwordHidden
                 )
-                Text(
-                    text = "Obligatorio *",
-                    textAlign = TextAlign.Start,
-                    fontSize = 12.sp,
-                    textDecoration = TextDecoration.None,
-                    letterSpacing = 0.sp,
-
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier
-                        .align(alignment = Alignment.Start)
-                        .padding(18.dp, 0.dp, 18.dp, 0.dp)
-
-                        .width(77.dp)
-
-                        //.height(18.dp)
-
-                        .alpha(1f),
-
-
-                    fontWeight = FontWeight.Medium,
-                    fontStyle = FontStyle.Normal,
-                    color = Color.Gray
-                )
-                PasswordTextField(
-                    value = viewModel.repetirContrasena,
-                    onValueChange = { newValue -> viewModel.onRepetirContrasenaChanged(newValue) },
-                    placeholder = "Repetir Contraseña",
-                    keyboardType = KeyboardType.Password,
-                    isPassword = true,
-                    passwordVisibility = true // Puedes controlar la visibilidad de la contraseña aquí
-                )
 
                 Button(
-                    onClick = { /* Realizar registro aquí */ },
+                    onClick = {viewModel.registerUser(viewModel.correo, viewModel.contrasena, navController)},
                     shape = RoundedCornerShape(topStart = 5.dp, topEnd = 5.dp, bottomStart = 5.dp, bottomEnd = 5.dp),
                     colors = ButtonDefaults.buttonColors(
                         backgroundColor = Color(red = 0.7216145992279053f, green = 0.015033637173473835f, blue = 0.015033637173473835f, alpha = 0.7900000214576721f)
@@ -225,7 +164,7 @@ fun RegistroUScreen() {
                         focusedElevation = 4.dp
                     )
                 ) {
-                    Text(text = "Guardar", color = Color.White)
+                    Text(text = "Continuar", color = Color.White)
                 }
             }
         }
